pyproject.toml
[project]
name = "stock-data-collect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask==3.0.0",
    "google-cloud-bigquery==3.14.1",
    "google-cloud-logging==3.9.0",
    "google-cloud-secret-manager==2.18.1",
    "gunicorn==21.2.0",
    "jpholiday==0.1.9",
    "pydantic==2.5.3",
    "pytest-cov==4.1.0",
    "pytest==7.4.4",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "tenacity==8.2.3",
    "pytest-mock==3.12.0",
    "black==23.12.1",
    "flake8==7.0.0",
    "mypy==1.8.0",
    "types-requests==2.31.0.10",
]
